# [generator] Use a lookup cache in SymbolTable to find types.