The way we use a `SymbolTable` is to fully populate it with types and then perform lots of lookups on it.  We can optimize this by building a lookup cache when we start accessing the types and then using it for all subsequent lookups.

If any additional types are added to the table, we clear the cache so it will be rebuilt.  In practice this doesn't happen often, so we are not constantly rebuilding the cache. (In testing the most we build the cache is twice, out of thousands of accesses.)

Scenario | Original | This PR
------------ | ------------- | -
jar2xml | 2645ms | 782ms
class-parse | 5563ms | 2555ms