@jonpryor 2019-07-26
If `SymbolTable` *itself* can be used from multiple threads -- and the locking within `EnsurePopulated()` implies that *maybe*, *perhaps* that can happen? -- then we need to ensure that `SymbolTable.symbols` and `SymbolTable.all_symbols_cache` are also protected.