@jpobst 2019-07-29
Added additional thread-safety guards.